feat : 관리자 모든 가게 조회 구현
- 관리자 모든 가게 조회 구현
- 활성화된 가게들 페이지네이션 처리해서 조회
- api로 불러올 때 type이 int여서 가격 부분은 BigDecimal -> int로 수정

+++ apimocking-main/apimocking-boilerplate/src/main/java/com/grepp/spring/app/model/admin/dto/AdminStoreMenuResponse.java
@@ -10,7 +10,7 @@ public record AdminStoreMenuResponse(
     String name,
 
     @Schema(description = "메뉴 가격", example = "4500")
-    BigDecimal price
+    int price
 
     ) {
